+++ src/raspberrypi/devices/scsihd_nec.cpp
@@ -137,26 +137,26 @@ vector<BYTE> SCSIHD_NEC::InquiryInternal() const
 	return HandleInquiry(device_type::DIRECT_ACCESS, scsi_level::SCSI_1_CCS, false);
 }
 
-void SCSIHD_NEC::AddErrorPage(map<int, vector<BYTE>>& pages, bool) const
+void SCSIHD_NEC::AddErrorPage(map<int, vector<byte>>& pages, bool) const
 {
-	vector<BYTE> buf(8);
+	vector<byte> buf(8);
 
 	// The retry count is 0, and the limit time uses the default value inside the device.
 
 	pages[1] = buf;
 }
 
-void SCSIHD_NEC::AddFormatPage(map<int, vector<BYTE>>& pages, bool changeable) const
+void SCSIHD_NEC::AddFormatPage(map<int, vector<byte>>& pages, bool changeable) const
 {
-	vector<BYTE> buf(24);
+	vector<byte> buf(24);
 
 	// Page can be saved
-	buf[0] = 0x80;
+	buf[0] = (byte)0x80;
 
 	// Make the number of bytes in the physical sector appear mutable (although it cannot actually be)
 	if (changeable) {
-		buf[0xc] = 0xff;
-		buf[0xd] = 0xff;
+		buf[0xc] = (byte)0xff;
+		buf[0xd] = (byte)0xff;
 
 		pages[3] = buf;
 
@@ -165,40 +165,40 @@ void SCSIHD_NEC::AddFormatPage(map<int, vector<BYTE>>& pages, bool changeable) c
 
 	if (IsReady()) {
 		// Set the number of tracks in one zone (PC-9801-55 seems to see this value)
-		buf[0x2] = (BYTE)(heads >> 8);
-		buf[0x3] = (BYTE)heads;
+		buf[0x2] = (byte)(heads >> 8);
+		buf[0x3] = (byte)heads;
 
 		// Set the number of sectors per track
-		buf[0xa] = (BYTE)(sectors >> 8);
-		buf[0xb] = (BYTE)sectors;
+		buf[0xa] = (byte)(sectors >> 8);
+		buf[0xb] = (byte)sectors;
 
 		// Set the number of bytes in the physical sector
 		int size = 1 << disk.size;
-		buf[0xc] = (BYTE)(size >> 8);
-		buf[0xd] = (BYTE)size;
+		buf[0xc] = (byte)(size >> 8);
+		buf[0xd] = (byte)size;
 	}
 
 	// Set removable attributes (remains of the old days)
 	if (IsRemovable()) {
-		buf[20] = 0x20;
+		buf[20] = (byte)0x20;
 	}
 
 	pages[3] = buf;
 }
 
-void SCSIHD_NEC::AddDrivePage(map<int, vector<BYTE>>& pages, bool changeable) const
+void SCSIHD_NEC::AddDrivePage(map<int, vector<byte>>& pages, bool changeable) const
 {
-	vector<BYTE> buf(20);
+	vector<byte> buf(20);
 
 	// No changeable area
 	if (!changeable && IsReady()) {
 		// Set the number of cylinders
-		buf[0x2] = (BYTE)(cylinders >> 16);
-		buf[0x3] = (BYTE)(cylinders >> 8);
-		buf[0x4] = (BYTE)cylinders;
+		buf[0x2] = (byte)(cylinders >> 16);
+		buf[0x3] = (byte)(cylinders >> 8);
+		buf[0x4] = (byte)cylinders;
 
 		// Set the number of heads
-		buf[0x5] = (BYTE)heads;
+		buf[0x5] = (byte)heads;
 	}
 
 	pages[4] = buf;
